backend/sync/requirements.txt
apify-client>=1.7.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
rapidfuzz>=3.0.0
pandas>=2.0.0
